Ensure calling Delete on a deleted file does not panic

+++ tsdb/engine/tsm1/tsm1.go
@@ -2017,6 +2017,11 @@ func (d *dataFile) Delete() error {
 	if err := d.close(); err != nil {
 		return err
 	}
+
+	if d.f == nil {
+		return nil
+	}
+
 	err := os.RemoveAll(d.f.Name())
 	if err != nil {
 		return err
